moved error-messages to configs/
