+++ app.js
@@ -10,7 +10,7 @@ const limiter = require('./configs/rate-limiter');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const { errorHandler } = require('./middlewares/error-handler');
 const NotFoundError = require('./errors/not-found-error');
-const { ROUTE_NOT_FOUND_MESSAGE } = require('./utils/error-messages');
+const { ROUTE_NOT_FOUND_MESSAGE } = require('./configs/error-messages');
 
 const app = express();
 const { PORT = 3000, MONGO_URL = 'mongodb://localhost:27017/newsdb' } = process.env;
